refactor(ffi): winscard: rename module buff_alloc -> buf_alloc
